Revert "Merge remote-tracking branch 'origin/HannesGren' into MartinBranch"
This reverts commit 35288ebf2e87c50e8ed58ea676a457ea1ff9b79a, reversing
changes made to 02f7bf1dff36f0bd57309b5ab1d83c5fb3cff0cb.

+++ Source/PROJ/DamageZone.h
@@ -33,7 +33,7 @@ private:
 
 	// The amount of damage the player(s) takes each time
 	UPROPERTY(EditAnywhere)
-	float DamageAmount = 1;
+	uint8 DamageAmount = 1;
 
 	// Whether or not the player dies instantly when entering the zone
 	UPROPERTY(EditAnywhere)
